Adds ProgressFileReader to take reading responsibility from ProgressFileWriter
